+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file needs to be updated with every significant pull request. It is used to
 
 ## Version 0.5.6
 * Expose request id to parent Rack application when using `ApplicationInsights::Rack::TrackRequest` middleware through `env['ApplicationInsights.request.id']`
+* Implement operation context functionality for `ApplicationInsights::Rack::TrackRequest`
 * Add functionality to accept a Request-Id header for telemetry correlation
 
 ## Version 0.5.5